Split out Mapping component. It was begging to be done. It's also bugging to be redesigned, but that's later.
