+++ src/component/window/url-mapping.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import NewMapping from '../mapping/new-mapping.js';
+import Mapping from '../mapping/mapping.js';
 
 const {func, array, string} = React.PropTypes;
 
@@ -15,39 +16,12 @@ export default class UrlMappingWindow extends React.Component {
     } = this.props;
 
     let {urlMappings} = this.props;
-    urlMappings = urlMappings.map((map, index) => {
-      const removeMapping = () => {
-        removeUrlMapping(map.url);
-      };
-
-      const toggleIsActive = () => {
-        toggleUrlMappingIsActive(map.url);
-      };
-
-      const isActiveClass = map.isActive ? 'on' : 'off';
-
-      return <li className="collection-item" key={index}>
-        <div>
-          <span className="col protocol">
-            http(s)://
-          </span>
-          <span className="col text-ellipsis mask">
-            {map.url}
-          </span>
-          <span className="seperator">
-            <i className="fa fa-chevron-right"></i>
-          </span>
-          <span className="col text-ellipsis new-url">
-            {map.newUrl}
-          </span>
-          <a href="#!" className="secondary-content" onClick={toggleIsActive}>
-            <i className={'fa fa-toggle-' + isActiveClass}></i>
-          </a>
-          <a href="#!" className="secondary-content" onClick={removeMapping}>
-            <i className="fa fa-remove"></i>
-          </a>
-        </div>
-      </li>;
+    urlMappings = urlMappings.map((mapping, index) => {
+      return <Mapping key={index}
+                mapping={mapping}
+                toggleActive={toggleUrlMappingIsActive}
+                remove={removeUrlMapping}
+              />;
     });
 
     return <div className="window url-mapping-window">
